+++ java_backend/src/main/java/com/example/dto/RoadMap.java
@@ -2,11 +2,6 @@ package com.example.model;
 
 import java.util.Date;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -16,8 +11,8 @@ public class RoadMap {
 	private String user_id;
 	private String title;
 	private String objective;
+	private String profile;
 	private Date deadline;
-
 	private Date created_at;
 	private Date updated_at;
 }
